Cleaning up CreditCard models
- removing cardInfo vars from main model
- splitting CreditCard into multiple files
