+++ FitpaySDK/Rest/Models/CreditCard/CreditCard.swift
@@ -20,12 +20,6 @@ import Foundation
     open var verificationMethods: [VerificationMethod]?
     open var externalTokenReference: String?
     open var info: CardInfo?
-    open var pan: String?
-    open var expMonth: Int?
-    open var expYear: Int?
-    open var cvv: String?
-    open var name: String?
-    open var address: Address?
     open var topOfWalletAPDUCommands: [APDUCommand]?
     open var tokenLastFour: String?
     
@@ -116,12 +110,6 @@ import Foundation
         case targetDeviceType
         case verificationMethods
         case externalTokenReference
-        case pan
-        case expMonth
-        case expYear
-        case cvv
-        case name
-        case address
         case topOfWalletAPDUCommands = "offlineSeActions.topOfWallet.apduCommands"
         case tokenLastFour
     }
@@ -148,12 +136,6 @@ import Foundation
         targetDeviceType = try? container.decode(.targetDeviceType)
         verificationMethods = try? container.decode(.verificationMethods)
         externalTokenReference = try? container.decode(.externalTokenReference)
-        pan = try? container.decode(.pan)
-        expMonth = try? container.decode(.expMonth)
-        expYear = try? container.decode(.expYear)
-        cvv = try? container.decode(.cvv)
-        name = try? container.decode(.name)
-        address = try? container.decode(.address)
         topOfWalletAPDUCommands = try? container.decode(.topOfWalletAPDUCommands)
         tokenLastFour = try? container.decode(.tokenLastFour)
     }
@@ -180,12 +162,6 @@ import Foundation
         try? container.encode(targetDeviceType, forKey: .targetDeviceType)
         try? container.encode(verificationMethods, forKey: .verificationMethods)
         try? container.encode(externalTokenReference, forKey: .externalTokenReference)
-        try? container.encode(pan, forKey: .pan)
-        try? container.encode(expMonth, forKey: .expMonth)
-        try? container.encode(expYear, forKey: .expYear)
-        try? container.encode(cvv, forKey: .cvv)
-        try? container.encode(name, forKey: .name)
-        try? container.encode(address, forKey: .address)
         try? container.encode(topOfWalletAPDUCommands, forKey: .topOfWalletAPDUCommands)
         try? container.encode(tokenLastFour, forKey: .tokenLastFour)
     }
@@ -384,300 +360,6 @@ import Foundation
         } else {
             completion(nil, ErrorResponse.clientUrlError(domain: CreditCard.self, client: client, url: url, resource: resource))
         }
-    }
-}
-
-open class CardMetadata: NSObject, ClientModel, Serializable {
-    
-    open var foregroundColor: String? // TODO: update to UIColor
-    open var issuerName: String?
-    open var shortDescription: String?
-    open var longDescription: String?
-    open var contactUrl: String?
-    open var contactPhone: String?
-    open var contactEmail: String?
-    open var termsAndConditionsUrl: String?
-    open var privacyPolicyUrl: String?
-    open var brandLogo: [Image]?
-    open var cardBackground: [Image]?
-    open var cardBackgroundCombined: [ImageWithOptions]?
-    open var cardBackgroundCombinedEmbossed: [ImageWithOptions]?
-    open var coBrandLogo: [Image]?
-    open var icon: [Image]?
-    open var issuerLogo: [Image]?
-    
-    private var _client: RestClient?
-    
-    var client: RestClient? {
-        get {
-            return self._client
-        }
-
-        set {
-            self._client = newValue
-
-            if let brandLogo = self.brandLogo {
-                for image in brandLogo {
-                    image.client = self.client
-                }
-            }
-
-            if let cardBackground = self.cardBackground {
-                for image in cardBackground {
-                    image.client = self.client
-                }
-            }
-
-            if let cardBackgroundCombined = self.cardBackgroundCombined {
-                for image in cardBackgroundCombined {
-                    image.client = self.client
-                }
-            }
-            
-            if let cardBackgroundCombinedEmbossed = self.cardBackgroundCombinedEmbossed {
-                for image in cardBackgroundCombinedEmbossed {
-                    image.client = self.client
-                }
-            }
-
-            if let coBrandLogo = self.coBrandLogo {
-                for image in coBrandLogo {
-                    image.client = self.client
-                }
-            }
-
-            if let icon = self.icon {
-                for image in icon {
-                    image.client = self.client
-                }
-            }
-
-            if let issuerLogo = self.issuerLogo {
-                for image in issuerLogo {
-                    image.client = self.client
-                }
-            }
-        }
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case foregroundColor
-        case issuerName
-        case shortDescription
-        case longDescription
-        case contactUrl
-        case contactPhone
-        case contactEmail
-        case termsAndConditionsUrl
-        case privacyPolicyUrl
-        case brandLogo
-        case cardBackground
-        case cardBackgroundCombined
-        case cardBackgroundCombinedEmbossed
-        case coBrandLogo
-        case icon
-        case issuerLogo
-    }
-
-    public required init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        
-        foregroundColor = try? container.decode(.foregroundColor)
-        issuerName = try? container.decode(.issuerName)
-        shortDescription = try? container.decode(.shortDescription)
-        longDescription = try? container.decode(.longDescription)
-        contactUrl = try? container.decode(.contactUrl)
-        contactPhone = try? container.decode(.contactPhone)
-        contactEmail = try? container.decode(.contactEmail)
-        termsAndConditionsUrl = try? container.decode(.termsAndConditionsUrl)
-        privacyPolicyUrl = try? container.decode(.privacyPolicyUrl)
-        brandLogo = try? container.decode(.brandLogo)
-        cardBackground = try? container.decode(.cardBackground)
-        cardBackgroundCombined = try? container.decode(.cardBackgroundCombined)
-        cardBackgroundCombinedEmbossed = try? container.decode(.cardBackgroundCombinedEmbossed)
-        coBrandLogo = try? container.decode(.coBrandLogo)
-        icon = try? container.decode(.icon)
-        issuerLogo = try? container.decode(.issuerLogo)
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        
-        try? container.encode(foregroundColor, forKey: .foregroundColor)
-        try? container.encode(issuerName, forKey: .issuerName)
-        try? container.encode(shortDescription, forKey: .shortDescription)
-        try? container.encode(longDescription, forKey: .longDescription)
-        try? container.encode(contactUrl, forKey: .contactUrl)
-        try? container.encode(contactPhone, forKey: .contactPhone)
-        try? container.encode(contactEmail, forKey: .contactEmail)
-        try? container.encode(termsAndConditionsUrl, forKey: .termsAndConditionsUrl)
-        try? container.encode(privacyPolicyUrl, forKey: .privacyPolicyUrl)
-        try? container.encode(brandLogo, forKey: .brandLogo)
-        try? container.encode(cardBackground, forKey: .cardBackground)
-        try? container.encode(cardBackgroundCombined, forKey: .cardBackgroundCombined)
-        try? container.encode(cardBackgroundCombinedEmbossed, forKey: .cardBackgroundCombinedEmbossed)
-        try? container.encode(coBrandLogo, forKey: .coBrandLogo)
-        try? container.encode(icon, forKey: .icon)
-        try? container.encode(issuerLogo, forKey: .issuerLogo)
-    }
-}
-
-open class TermsAssetReferences: NSObject, ClientModel, Serializable, AssetRetrivable {
-    open var mimeType: String?
-    
-    var client: RestClient?
-    var links: [ResourceLink]?
-
-    private static let selfResourceKey = "self"
-
-    private enum CodingKeys: String, CodingKey {
-        case links = "_links"
-        case mimeType
-    }
-
-    public required init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        
-        links = try container.decode(.links, transformer: ResourceLinkTypeTransform())
-        mimeType = try? container.decode(.mimeType)
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        
-        try? container.encode(links, forKey: .links, transformer: ResourceLinkTypeTransform())
-        try? container.encode(mimeType, forKey: .mimeType)
-    }
-
-    @objc open func retrieveAsset(_ completion: @escaping RestClient.AssetsHandler) {
-        let resource = TermsAssetReferences.selfResourceKey
-        let url = self.links?.url(resource)
-        if let url = url, let client = self.client {
-            client.assets(url, completion: completion)
-        } else {
-            let error = ErrorResponse.clientUrlError(domain: TermsAssetReferences.self, client: client, url: url, resource: resource)
-            completion(nil, error)
-        }
-    }
-}
-
-open class DeviceRelationships: NSObject, ClientModel, Serializable {
-    
-    open var deviceType: String?
-    open var deviceIdentifier: String?
-    open var manufacturerName: String?
-    open var deviceName: String?
-    open var serialNumber: String?
-    open var modelNumber: String?
-    open var hardwareRevision: String?
-    open var firmwareRevision: String?
-    open var softwareRevision: String?
-    open var created: String?
-    open var createdEpoch: TimeInterval?
-    open var osName: String?
-    open var systemId: String?
-
-    var client: RestClient?
-    var links: [ResourceLink]?
-
-    private static let selfResourceKey = "self"
-
-    private enum CodingKeys: String, CodingKey {
-        case deviceType
-        case links = "_links"
-        case deviceIdentifier
-        case manufacturerName
-        case deviceName
-        case serialNumber
-        case modelNumber
-        case hardwareRevision
-        case firmwareRevision
-        case softwareRevision
-        case created = "createdTs"
-        case createdEpoch = "createdTsEpoch"
-        case osName
-        case systemId
-    }
-
-    public required init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        
-        links = try container.decode(.links, transformer: ResourceLinkTypeTransform())
-        deviceType = try? container.decode(.deviceType)
-        deviceIdentifier = try? container.decode(.deviceIdentifier)
-        manufacturerName = try? container.decode(.manufacturerName)
-        deviceName = try? container.decode(.deviceName)
-        serialNumber = try? container.decode(.serialNumber)
-        modelNumber = try? container.decode(.modelNumber)
-        hardwareRevision = try? container.decode(.hardwareRevision)
-        firmwareRevision =  try? container.decode(.firmwareRevision)
-        softwareRevision = try? container.decode(.softwareRevision)
-        created = try? container.decode(.created)
-        createdEpoch = try container.decode(.createdEpoch, transformer: NSTimeIntervalTypeTransform())
-        osName = try? container.decode(.osName)
-        systemId = try? container.decode(.systemId)
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        
-        try? container.encode(links, forKey: .links, transformer: ResourceLinkTypeTransform())
-        try? container.encode(deviceType, forKey: .deviceType)
-        try? container.encode(deviceIdentifier, forKey: .deviceIdentifier)
-        try? container.encode(manufacturerName, forKey: .manufacturerName)
-        try? container.encode(deviceName, forKey: .deviceName)
-        try? container.encode(serialNumber, forKey: .serialNumber)
-        try? container.encode(modelNumber, forKey: .modelNumber)
-        try? container.encode(hardwareRevision, forKey: .hardwareRevision)
-        try? container.encode(firmwareRevision, forKey: .firmwareRevision)
-        try? container.encode(softwareRevision, forKey: .softwareRevision)
-        try? container.encode(created, forKey: .created)
-        try? container.encode(createdEpoch, forKey: .createdEpoch, transformer: NSTimeIntervalTypeTransform())
-        try? container.encode(osName, forKey: .osName)
-        try? container.encode(systemId, forKey: .systemId)
-    }
-
-    @objc func relationship(_ completion: @escaping RestClient.RelationshipHandler) {
-        let resource = DeviceRelationships.selfResourceKey
-        let url = self.links?.url(resource)
-        if let url = url, let client = self.client {
-            client.relationship(url, completion: completion)
-        } else {
-            completion(nil, ErrorResponse.clientUrlError(domain: DeviceRelationships.self, client: client, url: url, resource: resource))
-        }
-    }
-}
-
-open class CardInfo: Serializable {
-    open var pan: String?
-    open var expMonth: Int?
-    open var expYear: Int?
-    open var cvv: String?
-    open var creditCardId: String?
-    open var name: String?
-    open var address: Address?
-
-    private enum CodingKeys: String, CodingKey {
-        case pan
-        case creditCardId
-        case expMonth
-        case expYear
-        case cvv
-        case name
-        case address
-    }
-
-    public required init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        
-        pan = try? container.decode(.pan)
-        creditCardId = try? container.decode(.creditCardId)
-        expMonth = try? container.decode(.expMonth)
-        expYear = try? container.decode(.expYear)
-        cvv = try? container.decode(.cvv)
-        name = try? container.decode(.name)
-        address = try? container.decode(.address)
-        name = try? container.decode(.name)
     }
 
 }
